requirements.txt
accelerate==1.1.1
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
blendmodes==2024.1.1
blinker==1.9.0
Bottleneck==1.4.2
certifi==2024.8.30
charset-normalizer==3.4.0
clean-fid==0.1.35
click==8.1.7
clip-interrogator==0.6.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deepdanbooru==1.0.3
diskcache==5.6.3
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
einops==0.8.0
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
filterpy==1.4.5
Flask==3.1.0
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
gradio==3.41.2
gradio_client==0.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
imageio==2.36.1
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonmerge==1.9.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lark==1.2.2
lazy_loader==0.4
lightning-utilities==0.11.9
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mpmath==1.3.0
multidict==6.1.0
mysql-connector-python==9.1.0
narwhals==1.15.2
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
open_clip_torch==2.29.0
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
piexif==1.1.3
pillow==10.4.0
pillow-avif-plugin==1.4.3
propcache==0.2.1
protobuf==3.20.0
psutil==6.1.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
resize-right==0.0.2
rpds-py==0.22.0
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
sympy==1.13.1
tifffile==2024.9.20
timm==1.0.11
tokenizers==0.13.3
tomesd==0.1.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchdiffeq==0.2.5
torchmetrics==1.6.0
torchsde==0.2.6
torchvision==0.20.1+cu121
tqdm==4.67.1
trampoline==0.1.2
transformers==4.30.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
yarl==1.18.3
